requirements.txt
langchain>=0.2.0
langchain-openai>=0.0.5
langchain-core>=0.2.0
chromadb>=0.4.22
python-dotenv>=1.0.0
openai>=1.12.0
PyQt6>=6.6.1
PyPDF2>=3.0.1
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
python-multipart>=0.0.6
pandas>=2.0.0
yfinance>=0.2.30
numpy>=1.24.0
tiktoken>=0.5.1
